Split out steps into separate files
Split out steps into separate files
Added local fabric tests
removed integration tests which are now not needed
updated vscode ignore file so cucumber tests aren't packaged

contributes to #695
