+++ client/.vscode/launch.json
@@ -50,36 +50,6 @@
       }
     },
     {
-      "name": "Integration Tests",
-      "type": "extensionHost",
-      "request": "launch",
-      "runtimeExecutable": "${execPath}",
-      "args": [
-        "${workspaceFolder}/integrationTest/data/integrationTest.code-workspace",
-        "--extensionDevelopmentPath=${workspaceFolder}",
-        "--extensionTestsPath=${workspaceFolder}/out/integrationTest"
-      ],
-      "outFiles": [
-        "${workspaceFolder}/out/integrationTest/**/*.js"
-      ],
-      "preLaunchTask": "npm: watch"
-    },
-    {
-      "name": "Integration Tests (New Fabric)",
-      "type": "extensionHost",
-      "request": "launch",
-      "runtimeExecutable": "${execPath}",
-      "args": [
-        "${workspaceFolder}/integrationTest/data/integrationTest.code-workspace",
-        "--extensionDevelopmentPath=${workspaceFolder}",
-        "--extensionTestsPath=${workspaceFolder}/out/integrationTest"
-      ],
-      "outFiles": [
-        "${workspaceFolder}/out/integrationTest/**/*.js"
-      ],
-      "preLaunchTask": "npm: integrationTests"
-    },
-    {
       "name": "Node Integration Tests",
       "type": "extensionHost",
       "request": "launch",
