+++ client/cucumber/features/transaction.feature
@@ -4,7 +4,7 @@ Feature: Submit transaction
     Scenario Outline: Submit a transaction for a smart contract (local fabric)
         Given a <language> smart contract for <assetType> assets with the name <name> and version <version>
         And the Local Fabric is running
-        And connected to the 'Local Fabric' gateway
+        And I'm connected to the 'Local Fabric' gateway
         And the contract has been created
         And the contract has been packaged
         And the package has been installed
@@ -24,8 +24,8 @@ Feature: Submit transaction
         And the contract has been created
         And the contract has been packaged
         And the wallet 'myWallet' with identity 'conga' and mspid 'Org1MSP' exists
-        Given gateway 'myGateway' is created
-        And connected to the 'myGateway' gateway without association  
+        Given the gateway 'myGateway' is created
+        And I'm connected to the 'myGateway' gateway without association
         And the other fabric is setup with contract name <name> and version <version>
         When I submit the transaction 'createConga' with args '["Conga_001", "Big Conga"]'
         Then the logger should have been called with 'SUCCESS', 'Successfully submitted transaction' and 'No value returned from createConga'
