+++ client/cucumber/helpers/smartContractHelper.ts
@@ -16,11 +16,23 @@
 import * as vscode from 'vscode';
 import * as sinon from 'sinon';
 import * as path from 'path';
+import * as chai from 'chai';
+import * as sinonChai from 'sinon-chai';
+import * as chaiAsPromised from 'chai-as-promised';
 import { UserInputUtil } from '../../src/commands/UserInputUtil';
 import { UserInputUtilHelper } from './userInputUtilHelper';
 import { ExtensionCommands } from '../../ExtensionCommands';
 import { CommandUtil } from '../../src/util/CommandUtil';
 import { VSCodeBlockchainOutputAdapter } from '../../src/logging/VSCodeBlockchainOutputAdapter';
+import { PackageRegistry } from '../../src/packages/PackageRegistry';
+import { PackageRegistryEntry } from '../../src/packages/PackageRegistryEntry';
+import { BlockchainRuntimeExplorerProvider } from '../../src/explorer/runtimeOpsExplorer';
+
+import * as myExtension from '../../src/extension';
+
+chai.use(sinonChai);
+chai.use(chaiAsPromised);
+const should: Chai.Should = chai.should();
 
 export enum LanguageType {
     CHAINCODE = 'chaincode',
@@ -86,5 +98,145 @@ export class SmartContractHelper {
         }
 
         return contractDirectory;
+    }
+
+    public async packageSmartContract(name: string, version: string, language: string, directory: string): Promise<void> {
+        // Check that the package exists!
+        const _package: PackageRegistryEntry = await PackageRegistry.instance().get(name, version);
+        if (!_package) {
+            let workspaceFolder: vscode.WorkspaceFolder;
+
+            let workspaceFiles: vscode.Uri[];
+            if (language === 'JavaScript') {
+                workspaceFolder = { index: 0, name: name, uri: vscode.Uri.file(directory) };
+            } else if (language === 'TypeScript') {
+                workspaceFolder = { index: 0, name: name, uri: vscode.Uri.file(directory) };
+            } else if (language === 'Java') {
+                this.userInputUtilHelper.inputBoxStub.withArgs('Enter a name for your Java package').resolves(name);
+                this.userInputUtilHelper.inputBoxStub.withArgs('Enter a version for your Java package').resolves(version);
+                workspaceFolder = { index: 0, name: name, uri: vscode.Uri.file(directory) };
+            } else if (language === 'Go') {
+                this.userInputUtilHelper.inputBoxStub.withArgs('Enter a name for your Go package').resolves(name);
+                this.userInputUtilHelper.inputBoxStub.withArgs('Enter a version for your Go package').resolves(version);
+                workspaceFolder = { index: 0, name: name, uri: vscode.Uri.file(directory) };
+                workspaceFiles = [vscode.Uri.file('chaincode.go')];
+                this.userInputUtilHelper.findFilesStub.withArgs(new vscode.RelativePattern(workspaceFolder, '**/*.go'), null, 1).resolves(workspaceFiles);
+            } else {
+                throw new Error(`I do not know how to handle language ${language}`);
+            }
+
+            this.userInputUtilHelper.getWorkspaceFoldersStub.returns([workspaceFolder]);
+
+            await vscode.commands.executeCommand(ExtensionCommands.PACKAGE_SMART_CONTRACT, workspaceFolder, undefined, version);
+        }
+    }
+
+    public async installSmartContract(name: string, version: string): Promise<void> {
+        const blockchainRuntimeExplorerProvider: BlockchainRuntimeExplorerProvider = myExtension.getBlockchainRuntimeExplorerProvider();
+        const allTreeItems: any[] = await blockchainRuntimeExplorerProvider.getChildren();
+        const smartContracts: any[] = await blockchainRuntimeExplorerProvider.getChildren(allTreeItems[0]);
+        const installedContracts: any[] = await blockchainRuntimeExplorerProvider.getChildren(smartContracts[0]); // Installed smart contracts
+        const installedContract: any = installedContracts.find((contract: any) => {
+            return contract.label === `${name}@${version}`;
+        });
+
+        if (!installedContract) {
+            this.userInputUtilHelper.showPeersQuickPickStub.resolves(['peer0.org1.example.com']);
+            const _package: PackageRegistryEntry = await PackageRegistry.instance().get(name, version);
+
+            should.exist(_package);
+
+            this.userInputUtilHelper.showInstallableStub.resolves({
+                label: name,
+                data: {
+                    packageEntry: _package,
+                    workspace: undefined
+                }
+            });
+            await vscode.commands.executeCommand(ExtensionCommands.INSTALL_SMART_CONTRACT);
+        }
+    }
+
+    public async instantiateSmartContract(name: string, version: string, transaction: string, args: string, privateData: boolean): Promise<void> {
+        // Check if instantiated contract exists
+        const blockchainRuntimeExplorerProvider: BlockchainRuntimeExplorerProvider = myExtension.getBlockchainRuntimeExplorerProvider();
+        const allTreeItems: any[] = await blockchainRuntimeExplorerProvider.getChildren();
+        const smartContracts: any[] = await blockchainRuntimeExplorerProvider.getChildren(allTreeItems[0]);
+        const instantiatedContracts: any[] = await blockchainRuntimeExplorerProvider.getChildren(smartContracts[1]); // Installed smart contracts
+        const instantiatedContract: any = instantiatedContracts.find((contract: any) => {
+            return contract.label === `${name}@${version}`;
+        });
+
+        if (!instantiatedContract) {
+            this.userInputUtilHelper.showChannelStub.resolves({
+                label: 'mychannel',
+                data: ['peer0.org1.example.com']
+            });
+
+            const allPackages: Array<PackageRegistryEntry> = await PackageRegistry.instance().getAll();
+
+            const wantedPackage: PackageRegistryEntry = allPackages.find((packageEntry: PackageRegistryEntry) => {
+                return packageEntry.name === name && packageEntry.version === version;
+            });
+
+            this.userInputUtilHelper.showChaincodeAndVersionStub.resolves({
+                label: `${name}@${version}`,
+                description: 'Installed',
+                data: {
+                    packageEntry: wantedPackage,
+                    workspaceFolder: undefined,
+                }
+            });
+
+            this.userInputUtilHelper.inputBoxStub.withArgs('optional: What function do you want to call?').resolves(transaction);
+            this.userInputUtilHelper.inputBoxStub.withArgs('optional: What are the arguments to the function, (comma seperated)').resolves(args);
+
+            this.userInputUtilHelper.showYesNoQuickPick.resolves(UserInputUtil.NO);
+            if (privateData) {
+                this.userInputUtilHelper.showYesNoQuickPick.resolves(UserInputUtil.YES);
+                const collectionPath: string = path.join(__dirname, '../../integrationTest/data/collection.json');
+                this.userInputUtilHelper.browseStub.resolves(collectionPath);
+            }
+            await vscode.commands.executeCommand(ExtensionCommands.INSTANTIATE_SMART_CONTRACT);
+        }
+    }
+
+    public async upgradeSmartContract(name: string, version: string, transaction: string, args: string, privateData: boolean): Promise<void> {
+        this.userInputUtilHelper.showChannelStub.resolves({
+            label: 'mychannel',
+            data: ['peer0.org1.example.com']
+        });
+
+        const allPackages: Array<PackageRegistryEntry> = await PackageRegistry.instance().getAll();
+
+        const wantedPackage: PackageRegistryEntry = allPackages.find((packageEntry: PackageRegistryEntry) => {
+            return packageEntry.name === name && packageEntry.version === version;
+        });
+
+        this.userInputUtilHelper.showChaincodeAndVersionStub.resolves({
+            label: `${name}@${version}`,
+            description: 'Installed',
+            data: {
+                packageEntry: wantedPackage,
+                workspaceFolder: undefined,
+            }
+        });
+
+        // Upgrade from instantiated contract at version 0.0.1
+        this.userInputUtilHelper.showRuntimeInstantiatedSmartContractsStub.resolves({
+            label: `${name}@0.0.1`,
+            data: { name: name, channel: 'mychannel', version: '0.0.1' }
+        });
+
+        this.userInputUtilHelper.inputBoxStub.withArgs('optional: What function do you want to call?').resolves(transaction);
+        this.userInputUtilHelper.inputBoxStub.withArgs('optional: What are the arguments to the function, (e.g. ["arg1", "arg2"])', '[]').resolves(args);
+
+        this.userInputUtilHelper.showYesNoQuickPick.resolves(UserInputUtil.NO);
+        if (privateData) {
+            this.userInputUtilHelper.showYesNoQuickPick.resolves(UserInputUtil.YES);
+            const collectionPath: string = path.join(__dirname, '../../integrationTest/data/collection.json');
+            this.userInputUtilHelper.browseStub.resolves(collectionPath);
+        }
+        await vscode.commands.executeCommand(ExtensionCommands.UPGRADE_SMART_CONTRACT);
     }
 }
