+++ client/cucumber/steps/index.ts
@@ -14,10 +14,23 @@
 
 'use strict';
 
-// tslint:disable-next-line:no-var-requires
+// tslint:disable:no-var-requires
 const steps: any = require('./steps');
-
+const hooks: any = require('./hooks');
+const smartContractSteps: any = require('./smartContract.steps');
+const packageContractSteps: any = require('./package.steps');
+const installInstantiateUpgradeSteps: any = require('./installInstantiateUpgrade.steps');
+const wallletAndIdentitySteps: any = require('./walletAndIdentity.steps');
+const gatewaySteps: any = require('./gateway.steps');
+const localFabricSteps: any = require('./localFabric.steps');
 // We can break up step definitions into multiple files
 module.exports = function(): any {
+    hooks.call(this);
     steps.call(this);
+    smartContractSteps.call(this);
+    packageContractSteps.call(this);
+    installInstantiateUpgradeSteps.call(this);
+    wallletAndIdentitySteps.call(this);
+    gatewaySteps.call(this);
+    localFabricSteps.call(this);
 };
